pyproject.toml
[project]
name = "timeflies"
version = "0.2.0"
description = "Machine Learning for Aging Analysis in Drosophila Single-Cell Data"
authors = [{name = "Singh Lab"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorflow>=2.13.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scanpy>=1.9.0",
    "anndata>=0.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "shap>=0.42.0",
    "pyyaml>=6.0",
    "dill>=0.3.0",
    "xgboost>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.8.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
batch-correction = [
    "scvi-tools>=1.0.0",
    "torch>=2.0.0",
]

[project.scripts]
timeflies = "timeflies.cli.main:main_cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | build
    | dist
    | legacy
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "functional: Functional tests",
    "system: System tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
data_file = "coverage/.coverage"
omit = [
    "*/tests/*",
    "*/legacy/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.coverage.html]
directory = "coverage/html"

[tool.coverage.xml]
output = "coverage/coverage.xml"
